Pull Unicode segmentation support out of string and rewrite.
This is a nearly complete reimplementation, with a simpler and faster DFA,
providing a fold-like interface.

The describing rules are nearly identical to those in UAX #29, which makes them
much easier to write and understand.  Also, there's a debugging feature that
shows how the DFA evolves for a given string.