Major rewrite of library implementation.
New implementation supports an alternative syntax that's more expressive than
the one in R7RS, making it easier to handle complicated import/export situations
without explicit lists of names.  This was prompted by the SRFI 140 libraries,
which can't be simply expressed in the R7RS syntax.

The new syntax uses distinct keywords (define-library* and import*), keeping the
R7RS keywords as previously defined.  Regardless of the syntax, the internal
representation of libraries is the same, allowing the programmer to use
whichever is more convenient without worrying about compatibility.  (Caveat: the
R7RS include-library-declarations keyword can't be used to include a file using
define-library*.)

The new syntax introduces a couple of new concepts:

* "namesets" are just sets of names that can be combined with the usual set
  combinators, as well as filtered and renamed.  The leaf namesets are either
  explicit names, or the set of exported names from a specified library.

* Filtering and renaming using SRFI 115 regular expressions.  This is in
  addition to nameset intersection and simple one-to-one renames.

Note that the format of the metadata in compiled (.bin/.com) files has changed,
so it is necessary to re-compile any R7RS library/program files.  The loader
should detect and ignore compiled files with the old format, but if the
corresponding source file isn't available, the library will be unavailable.