Move most of the weak-pair code into a centralized place.
This makes it easier to maintain the weak-pair codebase and reduces the
possibility of introducing bugs by editing one use and not another.

In addition, this commit adds some useful generator support, which is compatible
with SRFI 158.  The generators simplify the implementation of procedures that
manipulate multiple weak lists, by hiding the elision of gc-reclaimed objects.