Workaround for cross-compiling with old microcode.
Otherwise toolchain doesn't work since it's looking for gc-claimed-object but
the microcode is using #f.

This is a kludge.  It deliberately ignores the case where #f is stored in a weak
car, which shouldn't be an issue for cross compiling.  Once the cross
compilation produces its final output, this won't matter since it will be using
gc-claimed-object.