Rework the interfaces of many weak-pair procedures to allow setters.
Most of the weak loops track the previous pair in a list (or alist) and dike out
elements that have been reclaimed.  But when the head of the list is reclaimed,
they can't do anything about that.  By adding optional set-X! arguments, those
operations now have a way to update the cell holding the list's head.

This not only allows the cleaning to be more effective, it also means that the
weak-list-set abstraction can just reuse existing procedures rather than having
their own slightly different ones.