Rewrite compound-predicate support.
New implementation uses hash tables with weak-list keys.  It also restructures
the construction so that each constructor generates a unique set of predicates
that can't be confused with those of any other.  Previously the "operator" name
was used to determine whether two compound predicates were related, which is not
sound.

However, using weak lists is not sufficient, since the datum carries strong
pointers to the elements of the key.  Either the datum will have to carry them
weakly, or I will have to implement ephemeral-list keys.