Fix bug: copy-library-db was preserving mutable data.
Specifically, it kept the environments in the source libraries, which could
cause problems.  The fixes guarantee that a library's properties are never
modified, only augmented by additional properties.  A pointer to the original
properties of the library is stashed, and when copying the library, it's
properties are restored to their original values.

The consequence of this change is that libraries may need to be re-loaded in the
copy, but since this is all automatic, that shouldn't be a problem.