Fix bugs in handling of library imports.
Since the import and export structures are identical, import expansions weren't
handling renames right, instead just passing along an export as an import.
Instead it is necessary to create a new import that uses the export's "to" name
as the import's "from" name.

This was trivial in the R7RS expansion process, but complicated in the MIT
expansion process.  So the MIT process was refactored to clarify what is going
on.