Add a chapter showing support for R7RS and various SRFIs.
All of the names that we implement are shown for each standard, and are indexed.
This should help people find documentation for names that aren't in the manual,
but are specified by an appropriate standard.  There's also some code to help
generate new entries in this chapter.

Also added a make target for generating all-on-one-page .html files.