Use -fno-builtin-floor if available.
This is because gcc7 miscompiles floor(0.5) to return -0 rather than
0 in some rounding modes.

We could conditionalize it on the version of gcc but that's more
work.  We could use a test program but test programs are annoying
because they don't work with cross-compilation.  It seems unlikely
that use of the compiler builtin floor is critical for performance,
so just keep it simple and disable it unconditionally.