Fix off-by-one error in unix socket path length.
The sun_path member of struct sockaddr_un need not be NUL-terminated;
it is a fixed-width field of the sort that strncpy was designed for.
Using a NUL terminator is only necessary for paths that don't fill
the sun_path member when we pass in the full structure size to bind,
as we do.