compiler/back: Save to temporaries if possible in CLEAR-REGISTERS!.
Make it update needed-registers while here -- this is necessary in
order for it to allocate temporaries as it goes, and helpful for most
callers anyway.

Update the documentation to reflect this and the 30-year-old change
to LOAD-MACHINE-REGISTER! so that it implies CLEAR-REGISTERS!.

This avoids unnecessary saves to memory homes and will help to
simplify the register allocator API, which I (and others, I suspect)
could never keep straight between clear/need/require/flush-registers.