Revert "compiler/back: Don't bother saving aliases for dead registers."
This reverts commit d248519275f7b9e406a5f09e4f882cc917192946.

This doesn't work after all, because the register might be needed by
_this_ instruction and might not have been copied to another alias or
saved to its home yet.

The issue I was trying to address with this change is that in

	(load-machine-register! source rax),

if source's only alias is rax and source is dead afterward, in our
zeal to clear rax we always allocate another alias for source (or
save it to its home) even if we're not actually going to use it again
later on.  We'll have to do that another way.