runtime: Stop blowing out the stack in list operations.
Exceptions: unzipN, unfold, pair-fold-right (not for any good reason,
I just got bored writing tests).

Note: This has the effect that internal continuations in many of the
higher-order procedures may be destructive even if the procedures
themselves are otherwise nondestructive.  No clear guidance in SRFI 1
about it.  However, it's just a matter of substituting reverse for
reverse! in all of these if this turns out to be a problem (but if
you do make this change, please write an automatic test that verifies
it).