Fix handling of dot token in reader.
The reader was allowing the dot token to appear in places where it is illegal.
For example, it could be used in vectors, where it was treated as a symbol,
which is explicitly forbidden by R7RS.

I changed the reader to never convert the dot token to a symbol, and to only be
allowed in the list-parsing context.  This revealed that there was an instance
of a quoted dot in another file, which again is forbidden.

This change may break any code that depends on dot being a symbol.