microcode: Fix bug in writing ucode stack traces to file.
In the v15 microcode merge, the type-abuse of casting a FILE * to
outf_channel was dropped, and thus so was the `C-c t' function's file
output at the console.

Do this is a slightly less abusive way that works with modern C and
modern platforms where enums are not necessarily wide enough, or
allowed by the semantics, to hold pointers.