A bunch of changes to get things to config/compile on Ubuntu 22.04 LTS.
Many of these changes were due to autoconf 2.71, which has obsoleted a bunch of
the macros we were using, and whose advice has considerably changed as to what
macros are needed for reasonably modern systems.

The rest are changes to the C files that were affected by the autoconf changes.