Handle circularity in arguments to equal?.
This is accomplished by detecting a pair of args that's been seen before and
returning the value from those args, or #t if there's no value yet.  Using a
true value allows the comparison to continue without committing to a particular
value for those args.  The actual value is determined when the original call
with those args finishes.