Refactor handling of scode for top-level define-syntax.
This introduces a new interface for manipulating the code for these syntax
definitions.  It's broken into two parts: scode-macro-definition for the
definition itself, and macro-transformer-expr for the transformer that goes in
the definition.

This cleans up the code a bit, but the real reason for the change is that I am
building some code that needs to assemble and disassemble these definitions.