Implement new kind of bundle.
The new style uses define-bundle-type to create a record whose fields
are the bundle's procedures.  The syntax shows the calling signature
of each procedure in the bundle; it also generates "invokers" that do
the indirect call fairly efficiently, at least compared to the old
style.

To gain a little more efficiency, the generated invokers could be
optimized in the same way that define-record-type does