Redesign of the new continuation parser.
Previous design was similar to existing parser in that it called
different parsers based on the frame type.  New design is built around
"tracked items", each of which has an updater that knows how to evolve
the item for each frame.  These updaters are much simpler than the
previous parsers, and they provide a good place to enforce invariants
on the value of the tracked item.

In addition, the control-point abstraction now provides the start and
end indexes of each frame, as well as the size of the control-point
that the frame is embedded in.  This information is needed by the
parser for various purposes.