Change handling of compiled-code stack frames.
Several of these stack frames are immediately followed by other
information that is actually a part of the frame.  The details of each
frame type are complex, but are now captured in the control-point
parser.

In particular, dynamic links are now handled correctly.

In addition, the parser now handles compiled entries other than return
addresses when they are encountered.  It's not obvious to me that this
can happen, but I can't rule it out.