Eliminate extra stack-frame layer and rename everything.
Instead there's now one kind of frame, which incorporates the
information from cpoint-frame and the tracked items from stack-frame*.
This eliminates indirect calls from stack-frame* to cpoint-frame.

The other consequence of this is that some stack-frame* operations
depended on stack-frame*/next, which isn't present in the new frame.
This is solved by changing those operations to take a stream of frames
instead.