Initial draft of new debugger.
It compiles and initial testing shows signs of life.

The primary change is twofold: to create a higher-level interface
for the debugger state; and to make the debugger state immutable.

It will be necessary to update all the debuggers for this new model,
but for now there's just an alternate version of the runtime's
debugger.