Build continuation-tree abstraction.
This is an immutable, stateless, lazy abstraction for navigating the
output of the continuation parser.  It is a replacement for the
recently introduced dstate abstraction, which had state that wasn't
used or needed by the Edwin debuggers.

I've tested the runtime debugger and the standard Edwin debugger and
they appear to be working OK.  I did no testing on artdebug, but since
it was already broken this shouldn't cause problems for anyone.