Revamp the representation of types.
Previously, a type had a name, which was either a symbol, or an
expression showing how the type was derived from other types.
However, the derivations are verbose and not very specific; for
example, NULL? printed as #[type NN (refinement constant)].

Now, each type has a name and a derivation.  The name is used in the
printed representation, unless the name is #F in which case the
derivation is shown.  NULL prints as #[type NN null] as it should.